requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
atlassian-python-api==3.41.19
attrs==25.1.0
Authlib==1.4.1
beautifulsoup4==4.13.3
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cryptography==44.0.0
cssselect2==0.7.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==7.3.2
email_validator==2.2.0
eval_type_backport==0.2.2
extractous==0.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
firebase-admin==6.6.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-google-firestore==0.5.0
langchain-google-vertexai==2.0.13
langchain-openai==0.3.4
langchain-postgres==0.0.14
langchain-text-splitters==0.3.6
langchain-unstructured==0.1.6
langgraph==0.2.70
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.7
libreoffice-convert==1.0
lxml==5.3.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
markdown-pdf==1.3.3
markdown2==2.5.3
markdown2slack==0.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
notion-client==2.3.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.61.1
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
pillow==11.1.0
primp==0.12.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.6
psycopg[binary]==3.2.6
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.24.6
PyMuPDFb==1.24.6
pyparsing==3.2.1
pypdf==5.4.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
schedule==1.2.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_bolt==1.22.0
slack_sdk==3.34.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
unstructured-client==0.32.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zopfli==0.2.3.post1
zstandard==0.23.0
